requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
tensorflow>=2.8.0
opencv-python>=4.5.3
pillow>=8.3.2 